--- example/src/Examples/Wallet/Header.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import { View } from "react-native";
-
-export const Header = () => {
-  return <View />;
-};
